[#2] Refactor: 파일 충돌해결 및 extension 합치기
